requirements.txt
# requirements.txt

flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
pandas
folium
Werkzeug==2.2.2
